compat with py 3.5
