+++ .gitignore
@@ -71,3 +71,4 @@ feeluown.egg-info/
 
 .pytest_cache/
 *.dat
+htmlcov/
